DDF-2906 Updated itests so most of them run in one container (#1817)
