+++ TestCatalog.java
@@ -20,14 +20,8 @@ import static org.codice.ddf.itests.common.catalog.CatalogTestCommons.ingest;
 import static org.codice.ddf.itests.common.catalog.CatalogTestCommons.ingestGeoJson;
 import static org.codice.ddf.itests.common.catalog.CatalogTestCommons.update;
 import static org.codice.ddf.itests.common.config.ConfigureTestCommons.configureEnforceValidityErrorsAndWarnings;
-import static org.codice.ddf.itests.common.config.ConfigureTestCommons.configureEnforcedMetacardValidators;
 import static org.codice.ddf.itests.common.config.ConfigureTestCommons.configureFilterInvalidMetacards;
-import static org.codice.ddf.itests.common.config.ConfigureTestCommons.configureMetacardValidityFilterPlugin;
 import static org.codice.ddf.itests.common.config.ConfigureTestCommons.configureShowInvalidMetacards;
-import static org.codice.ddf.itests.common.csw.CswQueryBuilder.NOT;
-import static org.codice.ddf.itests.common.csw.CswQueryBuilder.OR;
-import static org.codice.ddf.itests.common.csw.CswQueryBuilder.PROPERTY_IS_EQUAL_TO;
-import static org.codice.ddf.itests.common.csw.CswQueryBuilder.PROPERTY_IS_LIKE;
 import static org.codice.ddf.itests.common.csw.CswTestCommons.getCswInsertRequest;
 import static org.codice.ddf.itests.common.csw.CswTestCommons.getCswQuery;
 import static org.codice.ddf.itests.common.csw.CswTestCommons.getMetacardIdFromCswInsertResponse;
@@ -62,7 +56,6 @@ import java.time.OffsetDateTime;
 import java.time.ZoneId;
 import java.time.ZoneOffset;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.Date;
 import java.util.Dictionary;
 import java.util.HashMap;
@@ -91,7 +84,6 @@ import org.codice.ddf.itests.common.annotations.ConditionalIgnoreRule.Conditiona
 import org.codice.ddf.itests.common.annotations.SkipUnstableTest;
 import org.codice.ddf.itests.common.catalog.CatalogTestCommons;
 import org.codice.ddf.itests.common.config.UrlResourceReaderConfigurator;
-import org.codice.ddf.itests.common.csw.CswQueryBuilder;
 import org.codice.ddf.itests.common.utils.LoggingUtils;
 import org.codice.ddf.persistence.PersistentItem;
 import org.codice.ddf.persistence.PersistentStore;
@@ -108,7 +100,7 @@ import org.junit.rules.TestName;
 import org.junit.runner.RunWith;
 import org.ops4j.pax.exam.junit.PaxExam;
 import org.ops4j.pax.exam.spi.reactors.ExamReactorStrategy;
-import org.ops4j.pax.exam.spi.reactors.PerClass;
+import org.ops4j.pax.exam.spi.reactors.PerSuite;
 import org.osgi.service.cm.Configuration;
 import org.w3c.dom.Node;
 
@@ -125,13 +117,12 @@ import ddf.catalog.data.InjectableAttribute;
 import ddf.catalog.data.Metacard;
 import ddf.catalog.data.MetacardType;
 import ddf.catalog.data.types.Core;
-import ddf.catalog.data.types.Validation;
 
 /**
  * Tests the Catalog framework components. Includes helper methods at the Catalog level.
  */
 @RunWith(PaxExam.class)
-@ExamReactorStrategy(PerClass.class)
+@ExamReactorStrategy(PerSuite.class)
 public class TestCatalog extends AbstractIntegrationTest {
 
     private static final String METACARD_X_PATH = "/metacards/metacard[@id='%s']";
@@ -167,18 +158,8 @@ public class TestCatalog extends AbstractIntegrationTest {
     @BeforeExam
     public void beforeExam() throws Exception {
         try {
-            basePort = getBasePort();
-            getAdminConfig().setLogLevels();
-            getServiceManager().waitForRequiredApps(getDefaultRequiredApps());
-            getServiceManager().waitForAllBundles();
-            getCatalogBundle().waitForCatalogProvider();
-            getServiceManager().waitForHttpEndpoint(SERVICE_ROOT + "/catalog/query");
-            console = new KarafConsole(getServiceManager().getBundleContext(),
-                    features,
-                    sessionFactory);
-
-            configureRestForGuest();
-            getSecurityPolicy().waitForGuestAuthReady(REST_PATH.getUrl() + "?_wadl");
+            waitForSystemReady();
+            console = getConsole();
         } catch (Exception e) {
             LoggingUtils.failWithThrowableStacktrace(e, "Failed in @BeforeExam: ");
         }
@@ -1209,272 +1190,6 @@ public class TestCatalog extends AbstractIntegrationTest {
     }
 
     @Test
-    public void testEnforceValidityErrorsOnly() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Configure to enforce validator
-        configureEnforcedMetacardValidators(Collections.singletonList("sample-validator"),
-                getAdminConfig());
-
-        //Configure to enforce errors but not warnings
-        configureEnforceValidityErrorsAndWarnings("true", "false", getAdminConfig());
-
-        //Configure to show invalid metacards to show that invalid metacards were not ingested
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml");
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml",
-                false);
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard and warning metacard should be in results but not error one
-            response.body(containsString("warning metacard"));
-            response.body(containsString("clean metacard"));
-            response.body(not(containsString("error metacard")));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            deleteMetacard(id3, false);
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureEnforceValidityErrorsAndWarningsReset();
-            configureShowInvalidMetacardsReset();
-        }
-    }
-
-    @Test
-    public void testEnforceValidityWarningsOnly() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Configure to enforce validator
-        configureEnforcedMetacardValidators(Collections.singletonList("sample-validator"),
-                getAdminConfig());
-
-        //Configure to enforce warnings but not errors
-        configureEnforceValidityErrorsAndWarnings("false", "true", getAdminConfig());
-
-        //Configure to show invalid metacards to show that invalid metacards were not ingested
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml",
-                false);
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml");
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard and error metacard should be in results but not warning one
-            response.body(not(containsString("warning metacard")));
-            response.body(containsString("clean metacard"));
-            response.body(containsString("error metacard"));
-
-        } finally {
-            deleteMetacard(id1, false);
-            deleteMetacard(id2);
-            deleteMetacard(id3);
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureEnforceValidityErrorsAndWarningsReset();
-            configureShowInvalidMetacardsReset();
-        }
-    }
-
-    @Test
-    public void testEnforceValidityErrorsAndWarnings() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Configure to enforce validator
-        configureEnforcedMetacardValidators(Collections.singletonList("sample-validator"),
-                getAdminConfig());
-
-        //Configure to enforce errors and warnings
-        configureEnforceValidityErrorsAndWarnings("true", "true", getAdminConfig());
-
-        //Configure to show invalid metacards to show that invalid metacards were not ingested
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml",
-                false);
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml",
-                false);
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard should be in results but not invalid ones
-            response.body(not(containsString("warning metacard")));
-            response.body(containsString("clean metacard"));
-            response.body(not(containsString("error metacard")));
-
-        } finally {
-            deleteMetacard(id1, false);
-            deleteMetacard(id2);
-            deleteMetacard(id3, false);
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureEnforceValidityErrorsAndWarningsReset();
-            configureShowInvalidMetacardsReset();
-        }
-    }
-
-    @Test
-    public void testNoEnforceValidityErrorsOrWarnings() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Configure to enforce validator
-        configureEnforcedMetacardValidators(Collections.singletonList("sample-validator"),
-                getAdminConfig());
-
-        //Configure to enforce neither errors nor warnings
-        configureEnforceValidityErrorsAndWarnings("false", "false", getAdminConfig());
-
-        //Configure to show invalid metacards to show that invalid metacards were not ingested
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml");
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml");
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard should be in results but not invalid one
-            response.body(containsString("warning metacard"));
-            response.body(containsString("clean metacard"));
-            response.body(containsString("error metacard"));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            deleteMetacard(id3);
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureEnforceValidityErrorsAndWarningsReset();
-            configureShowInvalidMetacardsReset();
-        }
-    }
-
-    @Test
-    public void testQueryByErrorFailedValidators() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Don't enforce the validator, so that it will be marked but ingested
-        configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-
-        //Configure to show invalid metacards so the marked ones will appear in search
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        //Configure to not filter invalid metacards
-        configureFilterInvalidMetacards("false", "false", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml");
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml");
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    Validation.FAILED_VALIDATORS_ERRORS,
-                    "sample-validator")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard and warning metacard should be in results but not error one
-            response.body(not(containsString("warning metacard")));
-            response.body(not(containsString("clean metacard")));
-            response.body(containsString("error metacard"));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            deleteMetacard(id3);
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureShowInvalidMetacardsReset();
-            configureFilterInvalidMetacardsReset();
-        }
-    }
-
-    @Test
-    public void testQueryByWarningFailedValidators() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Don't enforce the validator, so that it will be marked but ingested
-        configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-
-        //Configure to show invalid metacards so the marked ones will appear in search
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        //Configure to not filter invalid metacards
-        configureFilterInvalidMetacards("false", "false", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml");
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml");
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    Validation.FAILED_VALIDATORS_WARNINGS,
-                    "sample-validator")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard and warning metacard should be in results but not error one
-            response.body(not(containsString("error metacard")));
-            response.body(not(containsString("clean metacard")));
-            response.body(containsString("warning metacard"));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            deleteMetacard(id3);
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureShowInvalidMetacardsReset();
-            configureFilterInvalidMetacardsReset();
-        }
-    }
-
-    @Test
     public void testFilterPlugin() throws Exception {
         // Ingest the metacard
         String id1 = ingestXmlFromResource("/metacard1.xml");
@@ -1512,201 +1227,6 @@ public class TestCatalog extends AbstractIntegrationTest {
             config.update(configProps);
             getServiceManager().stopFeature(true, "sample-filter");
             deleteMetacard(id1);
-        }
-    }
-
-    @Test
-    public void testFilterPluginWarningsOnly() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Configure not enforcing validators so invalid metacards can ingest
-        configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-
-        // Configure invalid filtering
-        configureMetacardValidityFilterPlugin(Arrays.asList("invalid-state=system-admin"),
-                getAdminConfig());
-
-        // Configure query to request invalid metacards
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        //Configure to filter metacards with validation warnings but not validation errors
-        configureFilterInvalidMetacards("false", "true", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml");
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml");
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard should be in results but not invalid one
-            response.body(not(containsString("warning metacard")));
-            response.body(containsString("clean metacard"));
-            response.body(containsString("error metacard"));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            deleteMetacard(id3);
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureFilterInvalidMetacardsReset();
-            configureMetacardValidityFilterPlugin(Arrays.asList(""), getAdminConfig());
-            configureShowInvalidMetacardsReset();
-        }
-    }
-
-    @Test
-    public void testFilterPluginErrorsOnly() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Configure not enforcing validators so invalid metacards can ingest
-        configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-
-        // Configure invalid filtering
-        configureMetacardValidityFilterPlugin(Arrays.asList("invalid-state=system-admin"),
-                getAdminConfig());
-
-        // Configure query to request invalid metacards
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        //Configure to filter metacards with validation errors but not validation warnings
-        configureFilterInvalidMetacards("true", "false", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml");
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml");
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard should be in results but not invalid one
-            response.body(not(containsString("error metacard")));
-            response.body(containsString("clean metacard"));
-            response.body(containsString("warning metacard"));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            deleteMetacard(id3);
-            configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureFilterInvalidMetacardsReset();
-            configureMetacardValidityFilterPlugin(Arrays.asList(""), getAdminConfig());
-            configureShowInvalidMetacardsReset();
-        }
-    }
-
-    @Test
-    public void testFilterPluginWarningsAndErrors() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Configure not enforcing validators so invalid metacards can ingest
-        configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-
-        // Configure invalid filtering
-        configureMetacardValidityFilterPlugin(Arrays.asList("invalid-state=system-admin"),
-                getAdminConfig());
-
-        // Configure query to request invalid metacards
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        //configure to filter both metacards with validation errors and validation warnings
-        configureFilterInvalidMetacards("true", "true", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml");
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml");
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard should be in results but not invalid one
-            response.body(not(containsString("error metacard")));
-            response.body(not(containsString("warning metacard")));
-            response.body(containsString("clean metacard"));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id3);
-            deleteMetacard(id2);
-            configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureFilterInvalidMetacardsReset();
-            configureMetacardValidityFilterPlugin(Arrays.asList(""), getAdminConfig());
-            configureShowInvalidMetacardsReset();
-        }
-    }
-
-    @Test
-    public void testFilterPluginNoFiltering() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        //Configure not enforcing validators so invalid metacards can ingest
-        configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-
-        // Configure invalid filtering
-        configureMetacardValidityFilterPlugin(Arrays.asList("invalid-state=system-admin"),
-                getAdminConfig());
-
-        // Configure query to request invalid metacards
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        //Configure to not filter metacard with validation errors or warnings
-        configureFilterInvalidMetacards("false", "false", getAdminConfig());
-
-        String id1 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleErrorMetacard.xml");
-        String id2 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleCleanMetacard.xml");
-        String id3 = ingestXmlFromResource(XML_RECORD_RESOURCE_PATH + "/sampleWarningMetacard.xml");
-
-        try {
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-
-            //clean metacard should be in results but not invalid one
-            response.body(containsString("error metacard"));
-            response.body(containsString("warning metacard"));
-            response.body(containsString("clean metacard"));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            deleteMetacard(id3);
-            configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureFilterInvalidMetacardsReset();
-            configureMetacardValidityFilterPlugin(Arrays.asList(""), getAdminConfig());
-            configureShowInvalidMetacardsReset();
         }
     }
 
@@ -2087,509 +1607,6 @@ public class TestCatalog extends AbstractIntegrationTest {
     @Test
     public void persistObjectToWorkspace() throws Exception {
         persistToWorkspace(100);
-    }
-
-    @Test
-    public void testValidationEnforced() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-        // Update metacardMarkerPlugin config with enforcedMetacardValidators
-        configureEnforcedMetacardValidators(Collections.singletonList("sample-validator"),
-                getAdminConfig());
-
-        String id1 = ingestXmlFromResource("/metacard1.xml");
-        String id2 = ingestXmlFromResource("/metacard2.xml", false);
-
-        try {
-            // Search for all entries, implicit "validation-warnings is null" and "validation-errors is null"
-            // should get added by ValidationQueryFactory
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 is in results AND not Metacard2
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            // Search for all entries that have no validation warnings or errors
-            query =
-                    new CswQueryBuilder().addPropertyIsNullAttributeFilter(Validation.VALIDATION_WARNINGS)
-                            .getQuery();
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 is in results AND not Metacard2
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            //Search for all entries that have validation-warnings from sample-validator or no validation warnings
-            //Only search that will actually return all entries
-
-            query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_EQUAL_TO,
-                    Validation.VALIDATION_WARNINGS,
-                    "*")
-                    .addPropertyIsNullAttributeFilter(Validation.VALIDATION_WARNINGS)
-                    .addLogicalOperator(OR)
-                    .getQuery();
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 and NOT metacard2 is in results
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            // Search for all metacards that have validation-warnings
-            query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_EQUAL_TO,
-                    Validation.VALIDATION_WARNINGS,
-                    "*")
-                    .getQuery();
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 and metacard2 are NOT in results
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1))));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-        } finally {
-            deleteMetacard(id1);
-            configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-            getServiceManager().stopFeature(true, "sample-validator");
-        }
-    }
-
-    @Test
-    public void testValidationUnenforced() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-        getServiceManager().stopFeature(true, "catalog-security-filter");
-        configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-
-        String id1 = ingestXmlFromResource("/metacard1.xml");
-        String id2 = ingestXmlFromResource("/metacard2.xml");
-
-        try {
-            // metacardMarkerPlugin has no enforcedMetacardValidators
-            // Search for all entries, implicit "validation-warnings is null" and "validation-errors is null"
-            // should get added by ValidationQueryFactory
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 is in results AND not Metacard2
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            // Search for all entries that have no validation warnings
-            query =
-                    new CswQueryBuilder().addPropertyIsNullAttributeFilter(Validation.VALIDATION_WARNINGS)
-                            .getQuery();
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 is in results AND not Metacard2
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            //Search for all entries that have validation-warnings or no validation warnings
-            //Only search that will actually return all entries
-            query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    Validation.VALIDATION_WARNINGS,
-                    "sampleWarnings")
-                    .addPropertyIsNullAttributeFilter(Validation.VALIDATION_WARNINGS)
-                    .addLogicalOperator(OR)
-                    .getQuery();
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 AND Metacard2 are in results
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2)));
-
-            // Search for all entries that are invalid
-            query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    Validation.VALIDATION_WARNINGS,
-                    "*")
-                    .getQuery();
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard2 is in results AND not Metacard1
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1))));
-
-            query =
-                    new CswQueryBuilder().addPropertyIsNullAttributeFilter(Validation.VALIDATION_WARNINGS)
-                            .addLogicalOperator(NOT)
-                            .getQuery();
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard2 is in results AND not Metacard1
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1))));
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            getServiceManager().stopFeature(true, "sample-validator");
-            getServiceManager().startFeature(true, "catalog-security-filter");
-        }
-    }
-
-    @Test
-    public void testValidationEnforcedUpdate() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-        // metacardMarkerPlugin has no enforced validators so both metacards can be ingested
-        final String id1 = ingestXmlFromResource("/metacard1.xml");
-        final String id2 = ingestXmlFromResource("/metacard2.xml");
-
-        try {
-            // Enforce the sample metacard validator
-            configureEnforcedMetacardValidators(Collections.singletonList("sample-validator"),
-                    getAdminConfig());
-
-            String metacard2Xml = getFileContent("metacard2.xml");
-            given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(metacard2Xml)
-                    .put(new DynamicUrl(REST_PATH, id1).getUrl())
-                    .then()
-                    .assertThat()
-                    .statusCode(HttpStatus.SC_BAD_REQUEST);
-
-            String metacard1Xml = getFileContent("metacard1.xml");
-            given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(metacard1Xml)
-                    .put(new DynamicUrl(REST_PATH, id2).getUrl())
-                    .then()
-                    .assertThat()
-                    .statusCode(HttpStatus.SC_OK);
-
-            String metacard1Path = format(METACARD_X_PATH, id1);
-            String metacard2Path = format(METACARD_X_PATH, id2);
-
-            executeOpenSearch("xml", "q=*").log()
-                    .all()
-                    .assertThat()
-                    .body(hasXPath(metacard1Path))
-                    .body(hasXPath(metacard1Path + "/string[@name='title']/value",
-                            is("Metacard-1")))
-                    .body(not(hasXPath(metacard1Path + "/string[@name='validation-errors']")))
-                    .body(not(hasXPath(metacard1Path + "/string[@name='validation-warnings']")))
-                    .body(hasXPath(metacard2Path))
-                    .body(hasXPath(metacard2Path + "/string[@name='title']/value",
-                            is("Metacard-1")))
-                    .body(not(hasXPath(metacard2Path + "/string[@name='validation-errors']")))
-                    .body(not(hasXPath(metacard2Path + "/string[@name='validation-warnings']")));
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            configureEnforcedMetacardValidators(Collections.singletonList(""), getAdminConfig());
-            getServiceManager().stopFeature(true, "sample-validator");
-        }
-    }
-
-    @Test
-    public void testValidationUnenforcedUpdate() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-        // metacardMarkerPlugin has no enforced validators so both metacards can be ingested
-        final String id1 = ingestXmlFromResource("/metacard1.xml");
-        final String id2 = ingestXmlFromResource("/metacard2.xml");
-
-        try {
-            String metacard2Xml = getFileContent("metacard2.xml");
-            given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(metacard2Xml)
-                    .put(new DynamicUrl(REST_PATH, id1).getUrl())
-                    .then()
-                    .assertThat()
-                    .statusCode(HttpStatus.SC_OK);
-
-            String metacard1Xml = getFileContent("metacard1.xml");
-            given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(metacard1Xml)
-                    .put(new DynamicUrl(REST_PATH, id2).getUrl())
-                    .then()
-                    .assertThat()
-                    .statusCode(HttpStatus.SC_OK);
-
-            configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-            String metacard1Path = format(METACARD_X_PATH, id1);
-            String metacard2Path = format(METACARD_X_PATH, id2);
-
-            executeOpenSearch("xml", "q=*").log()
-                    .all()
-                    .assertThat()
-                    .body(hasXPath(metacard1Path))
-                    .body(hasXPath(metacard1Path + "/string[@name='title']/value",
-                            is("Metacard-2")))
-                    .body(hasXPath(
-                            "count(" + metacard1Path + "/string[@name='validation-errors']/value)",
-                            is("1")))
-                    .body(hasXPath("count(" + metacard1Path
-                            + "/string[@name='validation-warnings']/value)", is("1")))
-                    .body(hasXPath(metacard2Path))
-                    .body(hasXPath(metacard2Path + "/string[@name='title']/value",
-                            is("Metacard-1")))
-                    .body(not(hasXPath(metacard2Path + "/string[@name='validation-errors']")))
-                    .body(not(hasXPath(metacard2Path + "/string[@name='validation-warnings']")));
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            configureShowInvalidMetacardsReset();
-            getServiceManager().stopFeature(true, "sample-validator");
-        }
-    }
-
-    @Test
-    public void testValidationFiltering() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        // Update metacardMarkerPlugin config with no enforcedMetacardValidators
-        configureEnforcedMetacardValidators(Arrays.asList(""), getAdminConfig());
-
-        // Configure invalid filtering
-        configureMetacardValidityFilterPlugin(Arrays.asList("invalid-state=system-admin"),
-                getAdminConfig());
-
-        // Configure the PDP
-        PdpProperties pdpProperties = new PdpProperties();
-        pdpProperties.put("matchOneMappings",
-                Arrays.asList(
-                        "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/role=invalid-state"));
-        Configuration config = configAdmin.getConfiguration("ddf.security.pdp.realm.AuthzRealm",
-                null);
-        Dictionary<String, ?> configProps = new Hashtable<>(pdpProperties);
-        config.update(configProps);
-
-        String id1 = ingestXmlFromResource("/metacard1.xml");
-        String id2 = ingestXmlFromResource("/metacard2.xml");
-        try {
-
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    Validation.VALIDATION_WARNINGS,
-                    "*")
-                    .addPropertyIsNullAttributeFilter(Validation.VALIDATION_WARNINGS)
-                    .addLogicalOperator(OR)
-                    .getQuery();
-
-            ValidatableResponse response = given().auth()
-                    .preemptive()
-                    .basic("admin", "admin")
-                    .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard2 is in results AND Metacard1
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2)));
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard2 is in results Metacard1
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            // Configure invalid filtering
-            configureMetacardValidityFilterPlugin(Arrays.asList("invalid-state=system-admin,guest"),
-                    getAdminConfig());
-
-            response = given().auth()
-                    .preemptive()
-                    .basic("admin", "admin")
-                    .header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard2 is in results AND Metacard1
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2)));
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard2 is in results Metacard1
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2)));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            getServiceManager().stopFeature(true, "sample-validator");
-            config = configAdmin.getConfiguration("ddf.security.pdp.realm.AuthzRealm", null);
-            configProps = new Hashtable<>(new PdpProperties());
-            config.update(configProps);
-        }
-    }
-
-    @Test
-    public void testValidationChecker() throws Exception {
-        getServiceManager().startFeature(true, "sample-validator");
-
-        configureEnforcedMetacardValidators(Arrays.asList(""), getAdminConfig());
-
-        // Configure invalid filtering
-        configureMetacardValidityFilterPlugin(Arrays.asList(""), getAdminConfig());
-
-        configureShowInvalidMetacards("true", "true", getAdminConfig());
-
-        String id1 = ingestXmlFromResource("/metacard1.xml");
-        String id2 = ingestXmlFromResource("/metacard2.xml");
-        try {
-
-            // Search for all entries, implicit "validation-warnings is null" and "validation-errors is null"
-            // should get added by ValidationQueryFactory
-            String query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_LIKE,
-                    "AnyText",
-                    "*")
-                    .getQuery();
-            ValidatableResponse response = given().header(HttpHeaders.CONTENT_TYPE,
-                    MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 is in results AND Metacard2 because showInvalidMetacards is true
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body((hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            // Search for all entries that have no validation warnings or errors
-            query =
-                    new CswQueryBuilder().addPropertyIsNullAttributeFilter(Validation.VALIDATION_WARNINGS)
-                            .getQuery();
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 is in results AND not Metacard2
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            //Search for all entries that have validation-warnings from sample-validator or no validation warnings
-
-            query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_EQUAL_TO,
-                    Validation.VALIDATION_WARNINGS,
-                    "*")
-                    .addPropertyIsNullAttributeFilter(Validation.VALIDATION_WARNINGS)
-                    .addLogicalOperator(OR)
-                    .getQuery();
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 and NOT metacard2 is in results
-            response.body(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1)));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-            // Search for all metacards that have validation-warnings
-            query = new CswQueryBuilder().addAttributeFilter(PROPERTY_IS_EQUAL_TO,
-                    Validation.VALIDATION_WARNINGS,
-                    "*")
-                    .getQuery();
-
-            response = given().header(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML)
-                    .body(query)
-                    .post(CSW_PATH.getUrl())
-                    .then();
-            // Assert Metacard1 and metacard2 are NOT in results
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id1))));
-            response.body(not(hasXPath(format(
-                    "/GetRecordsResponse/SearchResults/Record[identifier=\"%s\"]",
-                    id2))));
-
-        } finally {
-            deleteMetacard(id1);
-            deleteMetacard(id2);
-            getServiceManager().stopFeature(true, "sample-validator");
-            configureShowInvalidMetacardsReset();
-        }
     }
 
     private File copyFileToDefinitionsDir(String filename) throws IOException {
